+++ src/assets/static/js/wptelegram-widget--main.27168ff2.asset.php
@@ -1,0 +1,1 @@
+<?php return array('dependencies' => array('jquery', 'react', 'react-dom', 'wp-api-fetch', 'wp-i18n', 'wp-polyfill'), 'version' => '0e321655feaa778e1e66b7a49e1f31b6');